Purpose of this file:  - Handling of NaN values - Verifying if the data provided is adequate, and if not, add more relevant columns
Purpose of this file:

- Handling of NaN values
- Verifying if the data provided is adequate, and if not, add more relevant columns

Summary

NaN values -

Looking at the values, there seems to be an understanding that majority of the updated values are from "stronger" earthquakes, which are detectable. And all the activity that is not 
in detectable range will be supplied NULL values, but also any logical failures are listed as NULL.

It's not possible to determine the volume of "weaker" signals returning NULL, and logical error (Not with my current understanding of the calculations anyway). 
So it is safer to assume that majority of the signals are "weak", so I raised the floor from NULL to the lowest denominator across the board.

Country Level Data

I wanted to review an aggregated comparison across country/state level data, however, {place} value is not compatible for indexeing due to too much variability in the description/content.

To enable country level aggregation, I used the reverse geocoding method, via Nominatim, which takes long and lat column data and returns the location data, according to the zoom level. 
Zoom 3 was used to bring back country level data.

Manual sifting

America being a very large country, resulted in the country looking like an outlier, being massively populated with events. So I extracted state level data and integrated that into the dataset to enable 
a clearer comparison and zoom level of where the events are occuring.

(Don't run it, it takes 1-2 hours to run!)